Carousel add functions to move to a specific tile
When user clicks edit slide, we need to switch to the carousel view and edit a particular slide. If that slide is not visible in the carousel it would be a bad user experience. This commit adds functions goToTile(n) and goToActiveTile() so we can make sure an individual tile is visible when switching to edit that slide.