Carousel api changes
Added function to tell if a tile is visible (or partially hidden).
Added function to move a tile to the center of the viewport.
Added code when clicking a partially hidden tile to move it to the center.
Fixed some index problems - tiles are indexed starting at 1 for the first tile.