Carousel added dynamic-text linkages
When user enters new text for a slide, the label for the carousel tile and the grid view tile should automatically update to reflect the changes. Note there currently appears to be a bug where if you enter new text for the title it gets erased by the data-dynamic-text plugin.