Carousel slideshow update thumbnail and scroll up
If the user edits a slide and changes the image, the thumbnails in the grid view and carousel view should automatically update.
If user is in grid view and scrolled down, clicking edit on a tile will display the edit form; however, in case the window is scrolled down too far in some cases the edit form will not be visible and the user must scroll up to find it. To prevent this, if the user clicks edit on the tile, the page will scroll up if necessary to reveal the edit form.