Change the way we get image size to fix popup
There are some timing issues in the way we calculate the original image size. Changing it to use naturalWidth and adding a load handler in case the image is not finished loading yet.
Also fixed tabs in the popup version of the image editor.
Also fixed rich text editor in the popup version of the image editor.