Fixes for image cropping
Several changes here:
If image is rotated then remove all the size information and start over since we can't keep the same aspect ratios.
Added a check when polling for adjustments, stops polling if image gets removed from the DOM. This might happen if we are in a popup div that gets removed.
Modified the sizes to correctly handle a size that has width=0 or height=0. In that case the size is treated as an individual size, and user can resize the crop box freely (no aspect ratio).