Change resizing handles for hotspots
In the image editor, the hotspot boxes had three resize handles: left, right, and bottom/right.
The left and right handles adjusted the width of the box. The bottom/right handle let you resize the box but forced it to stay the same aspect ratio.
Changed to remove the left and right handles, and allow the bottom/right handle to freely resize the hotspot box.