Fix multiple rich text toolbars for text overlays
If an image had multiple text overlays, there was a separate text editor toolbar appearing for each one.
Modified the RTE code to fire rtefocus/rteblur events so I can tell when the user is inside the rich text editor.
Modified the image editor code to hide rich text editor toolbars when the rich text editor is not being used.
Moved rich text editor toolbar to the bottom of the image to keep it within the same tab while not interfering with the existing formatting.
Fixed a problem with the text overlay code that caused a javascript bug when loading a page with existing text overlays.