Fix upload files used with repeatable carousel
For a slideshow, upload files was broken after the recent changes to repeatable.js
The edit form was being moved to a different location, so after uploading new files the hidden inputs were not found.
Modified the code so the hidden inputs will be updated before  they are moved around.
Also modified the event handlers so the thumbnail image for the carousel will be updated for new images that are uploaded.