Fix for carousel left nav button
In the case where a few tiles were added, clicking the third or fourth tile would cause the active tile to be centered in the carousel viewport, and the first tile would be partially hidden. In that case if you clicked the left navigation button, the carousel failed to scroll back to show the first tile.
To fix this, adjusting the logic so the carousel will move even when the width of all tiles fit within the viewport.