Gallery thubmnail image bugfix
Fixed a problem where changing any image within a gallery item would update the thumbnail for the gallery.
Added a data-preview-field attribute to the repeatable object html, so the front-end could know which field provides the thumbnail.
Then when a change occurs within the gallery item, the front-end can update the thumbnail only if the expected field was updated.