Bugfix prevent editable placeholder clearing input
There are potential conflicts between dynamic placeholders and editable placeholders: the placeholder attributes on the input can sometimes change without warning.
Because of this, whenever we insert the placeholder value into the input, we must store that value for later comparison (because if we attempt to compare it against the placeholder value it might have been updated so it will not be what we expect and this might lead to erasing the input after the user enters a new value).