Modify popup to specify DOM location for popup
Previously the popup would always add to the end of the document body. Since this is outside the main form on the page, the content of the popup would not be submitted with the form. This commit changes the popup code so you can optionally specify the parent element where the popup should be added.