Hotspot display in popup first attempt
This is a first attempt at making the hotspot form appear in a popup.
In this case a single popup is used, and when the user edits the hotspot it is moved into the popup for editing, then moved back to the original position when the popup is closed.
This was needed because the other hotspot code expects the hotspot inputs to always be in a certain location; however, it is not an ideal implementation so I will be reworking it and modifying that other code as needed.