Repeatable thumbnail not updating for new item
For a repeatable form item, adding a new item was not updating the thumbnail image in the grid/gallery view. Note this bug was introduced due to another bugfix that prevented the thumbnail from updating for the wrong image.

Modified back-end code to add data-preview-field attribute to the template, so the front-end will know which field is used for the thumbnail.