Rich text editor 2 add custom styles from CMS
Dynamically load CSS rules for the custom styles that can be set in the CMS settings, so the styles will appear in the rich text editor.

Created a JSP file that outputs only the custom CMS style rules. When rich text editor initializes it loads the rules and adds them to the page.