Bugfix no add button in single input repeatable
For a "single input repeatable" such as SEO keywords, if you entered a value in the field but did not clicking the "add" button, your new keyword would not be added when you published the page. This commit modifies the javascript so it properly adds the ".toggle" hidden input so the keyword will be saved.