RTE2 prevent enhancement from being deleted
When a line is deleted, CodeMirror deletes all widgets attached to the line. This was causing any enhancements to be removed (with no way to "undo" the change). This commit adds code that detects when a line is deleted, then saves and restores the enhancement.

Also adds a clearHistory() call immediately after the content is loaded into the editor, so the user cannot undo past that.