RTE2 output raw html with data-rte-raw attribute
To accurately recognize "raw" html that the user has entered, adding an attribute "data-rte-raw" to all the HTML elements. Then when loading content into the rich text editor, that attribute can be detected to indicate it should remain as raw HTML, and it will not conflict with the HTML elements that the rich text editor is using.