RTE2 enhancement edit not working
On initially loading the rich text editor, if there is an enhancement, the Edit button was not working to edit the selected enhancement. However it was working after selecting the enhancement from the popup.

This commit adds some code that will set up the Edit link correctly when the enhancement is first added to the page.