RTE2 edit link on double click
CodeMirror doesn't handle double clicks very well, so initially links were implement so a single click on the link opened the Link Edit popup. However, that is causing usability problems and we want to require a double click to edit the link.
This commit adds code to simulate a double click event using mousedown events, so user can double click a link to open the Link Edit popup.