RTE2 newline handling for raw html
Modified HTML output and input to better support raw html.
For input: leaving newline characters in the input to force a new line
For output:
If a line ends with a character marked as raw HTML, do not add a <br> at the end of the line, instead add a newline character \n
For other lines, add a <br> to the end of the line
For lines that end with a block level element (like a UL or OL list) do not output a newline character.

This seems to prevent extra newlines from being added to the editor when importing arbitrary HTML for the most common cases. I don't think it will be possible to handle all cases perfectly, but this seems to be the best option.