Bugfix screens with left nav floating issue
On CMS screens that have a left nav (like settings pages), the previous change to the inputContainer styles is causing the first inputContainer to be strechted to the height of the left nav. Also when the page is scrolled and the left nav moves down, the inputContainer would stretch even more and prevent the page from scrolling correctly.

This commit adds some styles to the containing element to create a new "block formatting content" so when the float is cleared after the inputContainer it does not clear the left navigation float.