Richtext2 div import, carriage returns, etc
Enhancements to the rich text editor:

When importing HTML, treat \n as a special character. Display the newline as a carriage return symbol in the editor.
For output, convert back to \n

Added a keyboard shortcut Shift-Enter to manually create a newline.

Also made DIV elements import as raw HTML.

Also fixed a z-index issue where the toolbar submenu was being obscured if another rich text editor was below it.