RTE2 remove data-rte2-raw from HTML export
RTE2 was adding data-rte2-raw attribute to all exported raw HTML, to help it identify areas marked as raw HTML when the content was imported again.

This commit changes the export and import behavior so that will not be required:

Removed the data-rte2-raw attribute (by default, left it in for now as an option).

Modified the HTML import so it will convert HTML elements that exactly match the expected element and attribute. If a match is not found treat it as raw HTML. For example, <B> will be treated as bold, while <B CLASS="foo"> will be treated as raw HTML.

Added configuration parameters for styles so they can also accept any number of attributes, to be used for links since <A> element can have optional attributes.