RTE2 raw HTML to output BR element at end of line
Added a parameter (defaulting to true) so the end of each line in the editor will always output a BR element.

The previous behavior was to detect if the last character in a line was marked as raw HTML, and to output a newline character instead. This caused certain inconsistencies if the user entered text on an existing line then marked it as raw HTML (in some cases a BR element was changed to a newline in that case).

The new behavior has some downsides as well - if you enter multiple lines of text then mark them all as raw HTML, you could get some extra BR elements. To prevent this from happening, any HTML entered this way should be entered on a single line.