RTE2 bugfix for edit enhancement
The "edit" button for enhancements was not passing the 'reference' parameter (which is required to allow overrides for the enhancement).

According to the bug description: When using the `ToolUi.Referenceable` annotation, the developer has an option to pass in a class name. We use this on Images to allow an editor to override a caption/alt text and to add a link to a specific enhancement. When using the new RTE this data is no longer being saved in the CMS. Still works as expected in the old RTE.