RTE2 bugfix edit enhancement
The new rich text editor copied most of the logic from the old editor dealing with enhancements, however, there was some additional javascript inside enhancement.jsp that went behind the scenes to change things in the old editor.

This commit modified enhancement.jsp to provide the same functionality in the new rich text editor: when you click the "Edit" button on an enhancement, sometimes the enhancement will have a drop-down form for editing overrides to be used with the enhancement. After saving these overrides, they should be applied to the enhancment.