RTE temp fix for placeholder text overflowing
Currently the placeholder text is overlaid on top of the editor, but for multi-line placeholder this is not expanding the editor height, so the text is overflowing.
This is a temporary fix to prevent placeholder text from overflowing the editor by cropping the text to one line.
Another fix is in progress to properly implement editable placeholders, but it will take longer to implement so I'm rolling out this smaller change to make the editor usable in the meantime.