RTE add clipboardSanitizeFunction
Add a clipboardSanitizeFunction that can be set within the CMS.

For HTML from outside that is pasted into the editor, this function can modify the HTML before it is parsed by the editor. For example, to remove content.

To implement go to CMS settings > Debug > Extra Javacript, and add code like the following:

require(['jquery', 'v3/input/richtextCodeMirror'], function($, rte) {
    rte.clipboardSanitizeFunction = function($d) {
        // Example:
        // $d.find('br').after('<p>Adding another paragraph!</p>');
        return $d;
    };
});