RTE2 exclude track changes delete from word count
Modify the word count for rich text editor 2 to exclude areas that are marked DEL when using track changes.
Note this adds a bit of processing because now the word count has to convert the editor content to HTML and remove the deleted areas, instead of previous behavior getting just the text from the editor.