RTE2 bugfix enhancement focus and on click
Fix a problem introduced in a previous commit where clicking anywhere on the page was focusing on the first rich text editor on the page (due to behavior of popup close events).

Added ability to click anywhere within the enhancement to place cursor at the start of the line that contains the enhancement.