Bugfix for repeatable upload files
For repeatables such as slideshow, the "Upload Files" functionality was not working since we implemented the grid / gallery views. This occurred because the hidden inputs for each slide were being moved, and a callback function was no longer able to find them.

Moved the callback function so it would occur before the hidden inputs were moved.