Bugfix for rte toolbar buttons reloading frame
Adjust the jquery.frame.js code so it does not attempt to load url fragments (#name) into a new frame.

This is meant to fix an issue where a rich text editor was within a popup, and clicking the toolbar button was reloading the frame. The toolbar buttons are A elements with HREF="#".