RTE bugfix for enhancement changing line
If an enhancement is floated left or right, the HTML was created at the end of the line instead of the beginning. This resulted in the enhancement dropping to a lower line each time the page was published.

This commit removes some older code that had been refactored, so the enhancement HTML should always appear at the start of the line.