Bugfix sortables to handle nested elements
When a sortable list contains nested sortable lists of the same type, drag and drop was not possible.

This commit adds logic to allow drag and drop in the nested sortable (by performing a depth-first search of elements).

However, there still appears to be some faulty logic in positioning the drop placeholder, so the placeholder jumps around in some cases. This will be addressed in another commit.