RTE spell check display suggestions
For spelling errors, display suggestions to the user and let them choose a correction. Use right mouse button or Ctrl-Space to pop up the list of suggestions. Use keyboard arrows and press Enter to select a correction, or double-click with the mouse.

Does not play well with Track Changes yet.