Replace .throttle() function and add .debounce()
For better UI performance, we need to use .debounce() instead of .throttle().
This commit replaces the existing .throttle() function with an identical .throttle() and .debounce() code.
The result for the rich text editor should be better performance as the user is typing quickly we won't make certain updates until the user pauses in typing.