Bugfix for RTE and performance enhancements
There was some bad logic in the rich text editor using $.throttle at the wrong point, so the throttled function was being set in the object prototype. This lead to the change event firing for the wrong editor when there were mutiple RTEs on the page, and also caused the problem of the spellcheck sometimes not running when the editor initialized.

Also in some cases changed from using $.throttle() to using $.debounce(), to give better performance when the user starts typing.