Add a closed event to detect when popup is hidden
This is needed to more easily support RTE inline enhancements.
Currently the popup plugin triggers a "close" event on every click in the window, to tell all the popups to close.
In some cases that event has been used in an effort to determine when a particular popup will close; however, it is not well suited for that purpose because the event gets triggered on every click.
This commit adds a "closed" event that is triggered on an individual popup only when that popup is actually hidden, thus giving us a way to run some code only when our popup is closed.