RTE fix spacing of toolbar submenu items
A previous CSS change added a right margin to the last toolbar item, to leave room for the word count to be displayed. However, the rule was also applying to list items within toolbar submenus and causing those items to wrap in some cases.

This change makes the CSS more specific so it only applies to top-level toolbar items and not to submenu items.