RTE toolbar onclick always create new mark
Previously if you clicked the toolbar button to create a link or inline enhancemnet, it would check the editor to see if there was already a mark at the position you were editing. If there was a mark, it would edit the mark instead of creating a new mark. Howerver, this meant you could never create a nested element.
This commit changes the toolbar behavior so it will create a new mark, to allow for nested elements.