RTE adjustments to inline enhancements
Based on continuing discussion, modifying the way inline enhancements work.

Removes the inline enhancement style and toolbar button from default RTE - individual projects will be responsible for adding inline enhancement styles as needed.

Changes the style definition for inline enhancements so you must specify enhancementType to indicate the style is for an inline enhancement. Binding such a style to a toolbar icon will create a new enhancement and insert it in the document.

Note future commits are anticipated to make more changes such as removing the need for the toHTML and fromHTML configs, to make a more streamlined approach for custom HTML elements tied to enhancements.