RTE move rich text elements into toolbar submenu
For the RICH_TEXT_ELEMENTS inline enhancements, adding them individually to the toolbar will cause the toolbar to get too large. Instead adding them to an "Inline" toolbar submenu. Note we might want to change this in the future to allow the submenu name to be customized or multiple submenus to be used, but for now sticking with a single submenu.