RTE fix javscript error when table cell is null
There is a bug in handsontable that puts a null value in table cells when you create a new cell. This commit prevents a javascript error from occurring when you edit such a null cell. Note we are waiting on a potential fix to handsontable to fix the underlying issue of having null value in the first place: https://github.com/handsontable/handsontable/issues/2816