RTE bugfix readonly mode hide toolbars + no click
When the editor is in read only mode, it was possible to use the toolbar to modify the content. Also if there was a link in the content, it was possible to double click and modify the link. Also if there was an enhancement, it was possible to move, change, or remove the enhancement.
This commit removes the toolbar (from the editor and from enhancements), plus it prevents click events for links.