RTE support popup false for inline enhancement FE
This commit adds front-end code to support a popup:false parameter in the RICH_TEXT_ELEMENTS data, to prevent a popup form from appearing when user adds that inline enhancement. Note this commit only handles the front-end code, additional work will be needed to actually make the poup:false parameter appear in the RICH_TEXT_ELEMENTS variable that is placed on the page.