RTE remove double click handler
Previously we used double click to edit styles that have an onclick specification (such as links or inline enhancements). We have since changed the UI so placing the cursor within the style shows a dropdown with an edit button, so the double-click is no longer needed. This commit removes the double click code.