Nested popups close child should not close parent
In the case of nested popups, closing the child popup is also causing the parent popup to close. This is causing a problem in the rich text editor: when editing tables we popup another editor, which then can pop up a form to edit links. When the link popup is closed, it is automatically closes the parent popup as well, because the 'close' event propagates up through the DOM. The fix involves stopping the propagation so the close event does not bubble up to the parent.

# Conflicts:
#	tool-ui/src/main/webapp/script/v3/plugin/popup.js