RTE table track changes bugfix
Improve support for track changes within tables:

Fixed a bug where clearing the editor content was not necessarily clearing all marks. This had the effect of marking all the content in the table cell as "new" because a stray track-changes-insert mark was left in the editor.

Added style to the table display in the main editor, so track changes styles will appear in the table (before user clicks to edit the cell).

Added some logic when user clicks to edit the cell, on the popup editor maintain the same track-changes setting as the main editor. So if the user is marking changes in the main editor, then when editing a table cell track changes will still be on and changes will be marked there as well.