RTE prevent nested child popup from closing parent
In the case where one popup is nested within another, added some logic to the close event to prevent a child popup from closing the parent popup. We tried this previously by stopping event propagation, but that ended up breaking something else, so this solution still lets the event propagate.