RTE enhancement move prevent scroll soemtimes
Fixes the following problem:
In RTE create a table, then edit a table cell.
While editing the cell in a popup editor, add several lines, then add an enhancement, then move the enhancement up and down.
When you move the enhancement, note that behind the popup, the main window scroll position is changing, but the popup stays fixed.

This commit adds some logic to the scroll code, so it only scrolls the page when the editor is not position:fixed.