RTE table cell popup editor add save/cancel button
Add a heading to the popup table cell editor, plus add save/cancel buttons to make it more obvious how to save the table cell changes.
Note to prevent the user from typing content and accidentally losing it, should the popup be closed by other means (such as pressing escape or clicking outside the popup) the behavior in that case will still be to save the table cell.