Bugfix clicking popup content div closes popup
Clicking directly on the popup "content" div is causing the popup to close when it should not. The logic that detects clicks was checking if the click was on a child of the "content" div, but it was not accounting for clicks that were directly on the div. This commit adjusts the logic so clicks inside the div, or directly on the div, will be considering inside the popup and will not close the popup.