RTE remove rawCleanup for performance
To improve performance in the RTW when man raw marks are used, remove the rawCleanup() function. This also fixes an issue where a single element is split into multiple elements if raw marks exist within it. Note this could have a side effect of allowing the user to create invalid HTML if the user attempts to add styles to the middle of an area marked as raw HTML.