RTE add contextRoot option
Add a contextRoot option that can be set to a particular HTML element to use as the "root" when determining if other styles can be applied at the cursor position. For example, if a particular RTE will be used to edit content within a <HEADING> element, then you can set contextRoot="heading".