RTE bugfix single lines styles breaking lines
If a style has the singleLine parameter set to true, then that style should not be allowed to span across multiple lines.  This was already being enforced when the style is added; however, if the user edited the style and pressed Enter then the style was expanding to the second line. This commit ensure that the style is not duplicated on the second line (as this was causing a problem if the user then pressed backspace to remove the extra line, and we were ending up with two style markers).