RTE let block styles store attribute values
Added code so block styles in the editor will have a data object that can be used to save attribute values (and possibly other information).

Currently inline styles have a "mark" object in CodeMirror where additional data can be stored, but block styles are only represented in CodeMirror using a classname on the line; however, CodeMirror has a lineHandle object that always moves with the line, so we will use the lineHandle to save data for any of the line styles that are applied to the line.

The blockSetStyle() function has been modified to let you specify attribute values that should be saved along with the style.

Two additional functions blockSetLineData() and blockGetLineData() have been provided to let you set or retrieve the data based on a style and a line number.

Note additional work will need to be done so inline enhancments can have access to change the attributes on a line style.