RTE create getContext function with new logic
New function getContext to return a list of elements that are considered to be the context of the current cursor position or selection. Note you can have multiple elements in the context because user might have selected across multiple styles or lines, so when checking the context you must ensure that all of those contexts are allowed.