RTE bugfix position of table cell dropdown
When we updated to the most recent version of handsontable, it broke the way we position the dropdown menu. We are using a "non-standard" way to display the context menu since handsontable does not provide an official API for that, and apparently they changed the arguments used for the menu.open() function we are using.