RTE bugfix inline enhancements clear button
For inline enhancements, if you click the "Clear" button for characters in the middle of the enhancement, it will split the mark into two marks. We need to prevent that, so this commit modifies the clear function so it will not split inline enhancement marks. You can clear one side or the other, or clear the entire mark, but you cannot clear just a few characters in the middle of the mark.