Filters out ViewModel classes that aren't assignable to the view class.
Fixes issue where ViewModel#findViewModelClass could return null with a warning that multiple valid view model classes were found and therefore the result is ambiguous, when in fact certain view model classes were being included when they shouldn't be. Specifically, when a view class was passed as an argument, there was no check that the view model class and the view class were compatible. This commit adds that check.