Wrap jquery.dropdown in a requirejs define
The jquery.dropdown.js file is being loaded as a requirejs AMD module, but it does not define() itself as a module, even though it has a dependency and requires another module. This is causing timing issues in other code that is using dropdown as a dependency.

This commit removes the require() statement and instead wraps all the code in a define() using a dependency.