RTE add toggle style parameter for inline enh
For inline enhancements, add a toggle parameter to the style definition, so the toolbar button can act as a toggle to apply or remove the style. For example, if the style is a simple bold or italic element, you can set it to toggle.

Note this commit is the front-end code only, additional changes will need to be made for the backend to specify the toogle setting for each inline enhancement style.