RTE do not track changes in html mode
When switching from html mode to rich text mode, if track changes was previously turned on, all the HTML was marked as a change. This commit changes the behavior so when you switch from html mode to rich text mode, it temporarily turns off track changes so the text will not be marked as an insertion.