RTE bugfix track changes and clearing styles
The same part of code was causing a couple problems:

When selecting a single character, it was not possible to clear the style.

When track changes is on, and selecting a range of characters, typing or pasting should mark the selected range as a deletion, then add the new characters marked as an insertion. Instead it was marking the deletion but not adding the new text.

This commit fixes both issues.