RTE bugfix dropdown click after blur
Previous changes to hide the dropdown on blur caused a problem because clicking on the dropdown link (Edit/Clear) actually blurred the editor, which hid the dropdown, which prevented the click from being successful. This commit adds a timeout after the blur before hiding the dropdown, to give the click event time to register.