RTE change comment collapse to single comment
Currently the RTE toolbar button toggles collapse on all comments. This commit changes the behavior to toggle collapse only for those comments that are within the current range. Note it is still possible to a toggle all comments, by selecting all content in the editor (Ctrl-A or Cmd-A) then clicking the toolgar button.