Change event should ignore inputs without names
The "change" event handler looks for form inputs and textareas to determine if changes have been made; however, CodeMirror adds a few inputs and textareas to do its job, and these cause problems because it always appears that they have changed. This commit modifies the change event handler so it only looks for inputs and textareas that have a "name" attribute (and thus are submitted when the form is posted), so it will ignore the CodeMirror inputs.