RTE update the textarea and trigger change event
Changes to the RTE were not always updating the original textarea, and were not triggering a "change" event to indicate the field had changed. This commit periodically updates the textarea (even if the preview window is not showing), and when doing so triggers a change event so the colored change indicator shows up for the input. Note we debounce this agressively to limit how often we need to export the editor content into HTML.