RTE bugfix for change indicator when readonly
When the RTE is in readonly mode, the change indicator was sometimes appearing (in cases where the parsed HTML gets modified during parsing). This commit adds some code to ensure that the textarea is never updated when the editor is in readonly mode, to prevent parse changes from beingn submitted and changing the value of the texteara.