RTE bugfix toggle inline style across lines
When toggling an inline style with a range that goes across multiple lines, the style is being removed on the first line, but not on the other lines. I traced this to a side effect of triggering a change event for each line that is changed, which for some reason modified the other marks. I modified the code until a change event will not be triggered until all the lines are processed, and this seems to make the problem go away.