RTE add text to a mark after clearing part of mark
In the RTE, if you clear part of a mark, the RTE then lets you enter text outside the mark; but at that point you can never move the cursor to the right side of the mark and add more text to the mark.

This commit still lets you clear the mark and enter text outside the mark, but then if you return to the mark it should again be "inclusive right" and let you enter text inside the mark. This new behavior is modeled after other word processors.