RTE bugfix toolbar submenu cutting off on right
When the RTE toolbar submenu is displayed on the right side it is sometimes not wide enough and the links are cut off. This appears to be caused by the text wrapping due to spaces in the toolbar links. The wrapped text is hidden by overflow:hidden, so you never see it.
This commit fixes the problem by preventing the text from wrapping, and therefore the submenu stretches fully to show all the text.